requirements.txt
asgiref==3.6.0
boto3==1.26.51
cement==2.8.2
certifi==2022.12.7
charset-normalizer==2.0.12
colorama==0.4.3
Django==4.1.5
django-environ==0.9.0
django-extensions==3.2.1
django-filter==22.1
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
fake-factory==9999.9.9
Faker==16.3.0
future==0.16.0
idna==3.4
importlib-metadata==6.0.0
install==1.3.5
jmespath==0.10.0
Markdown==3.4.1
pathspec==0.9.0
Pillow==9.4.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.7
requests==2.26.0
s3transfer==0.6.0
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.3
termcolor==1.1.0
tzdata==2022.7
urllib3==1.26.14
wcwidth==0.1.9
zipp==3.11.0
asgiref==3.6.0
boto3==1.26.51
botocore==1.29.165
cement==2.8.2
certifi==2022.12.7
charset-normalizer==2.0.12
colorama==0.4.3
Django==4.1.5
django-environ==0.9.0
django-extensions==3.2.1
django-filter==22.1
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
fake-factory==9999.9.9
Faker==16.3.0
faker_education==1.2.1
future==0.16.0
idna==3.4
importlib-metadata==6.0.0
install==1.3.5
jmespath==0.10.0
Markdown==3.4.1
pathspec==0.9.0
Pillow==9.4.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.7
PyYAML==6.0.1
requests==2.26.0
s3transfer==0.6.0
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.3
termcolor==1.1.0
tzdata==2022.7
urllib3==1.26.14
wcwidth==0.1.9
zipp==3.11.0
